Refactor structure of the code
