--- src/main/java/com/eventscheduler/controller/Observer.java
@@ -1,11 +1,0 @@
-package com.eventscheduler.controller;
-
-/**
- * Observer interface for Observer Pattern
- */
-public interface Observer {
-    /**
-     * Updates the observer
-     */
-    void update();
-}
