+++ src/main/java/com/eventscheduler/controller/calendar/utils/DrawEventsUtils.java
@@ -1,9 +1,8 @@
-package com.eventscheduler.controller.utils;
+package com.eventscheduler.controller.calendar.utils;
 
-import com.eventscheduler.controller.CalendarController;
-import com.eventscheduler.model.EventModel;
+import com.eventscheduler.controller.calendar.CalendarController;
+import com.eventscheduler.model.dockument.EventModel;
 import javafx.scene.control.ScrollPane;
-import javafx.scene.layout.FlowPane;
 import javafx.scene.layout.VBox;
 import javafx.scene.text.Text;
 
@@ -16,7 +15,7 @@ public class DrawEventsUtils {
 
     private CalendarController calendarController;
 
-    public DrawEventsUtils(CalendarController calendarController){
+    public DrawEventsUtils(CalendarController calendarController) {
         this.calendarController = calendarController;
     }
