--- src/main/java/com/eventscheduler/model/Model.java
@@ -1,7 +1,0 @@
-package com.eventscheduler.model;
-
-/**
- * Model interface for Model-View-Controller Pattern
- */
-public interface Model {
-}
