+++ src/main/java/com/eventscheduler/model/dockument/Model.java
@@ -1,0 +1,7 @@
+package com.eventscheduler.model.dockument;
+
+/**
+ * Model interface for Model-View-Controller Pattern
+ */
+public interface Model {
+}
