+++ src/main/java/com/eventscheduler/observer/Observer.java
@@ -1,0 +1,11 @@
+package com.eventscheduler.observer;
+
+/**
+ * Observer interface for Observer Pattern
+ */
+public interface Observer {
+    /**
+     * Updates the observer
+     */
+    void update();
+}
