+++ src/main/java/module-info.java
@@ -10,6 +10,18 @@ module com.example.eventscheduler {
     exports com.eventscheduler;
     exports com.eventscheduler.model;
     opens com.eventscheduler.model to javafx.fxml;
-    exports com.eventscheduler.controller;
-    opens com.eventscheduler.controller to javafx.fxml;
+    exports com.eventscheduler.controller.calendar;
+    opens com.eventscheduler.controller.calendar to javafx.fxml;
+    exports com.eventscheduler.controller.event;
+    opens com.eventscheduler.controller.event to javafx.fxml;
+    opens com.eventscheduler.controller.event.utils to javafx.fxml;
+    exports com.eventscheduler.controller.event.utils;
+    exports com.eventscheduler.observer;
+    opens com.eventscheduler.observer to javafx.fxml;
+    exports com.eventscheduler.view;
+    opens com.eventscheduler.view to javafx.fxml;
+    exports com.eventscheduler.model.dockument;
+    opens com.eventscheduler.model.dockument to javafx.fxml;
+    exports com.eventscheduler.model.dao;
+    opens com.eventscheduler.model.dao to javafx.fxml;
 }
