+++ src/main/resources/com/eventscheduler/EventForm.fxml
@@ -13,7 +13,7 @@
 <?import javafx.scene.layout.RowConstraints?>
 <?import javafx.scene.text.Font?>
 
-<GridPane alignment="TOP_CENTER" hgap="10.0" prefHeight="400.0" prefWidth="600.0" vgap="10.0" xmlns="http://javafx.com/javafx/17" xmlns:fx="http://javafx.com/fxml/1" fx:controller="com.eventscheduler.controller.EventController">
+<GridPane alignment="TOP_CENTER" hgap="10.0" prefHeight="400.0" prefWidth="600.0" vgap="10.0" xmlns="http://javafx.com/javafx/17" xmlns:fx="http://javafx.com/fxml/1" fx:controller="com.eventscheduler.controller.event.EventController">
 
     <Label text="Title:" GridPane.columnIndex="0" GridPane.rowIndex="0">
       <font>
